feat : service interface를 추가
